requirements.txt
streamlit==1.37.1
pymongo==4.8.0
dnspython==2.6.1
pillow==10.4.0
bcrypt==4.2.0
pyarrow==16.1.0
numpy==1.26.4

openai==1.43.0
groq==0.11.0
toolhouse==0.4.3
python-dotenv==1.0.1
toml==0.10.2
